+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Each list item should be prefixed with `(patch)` or `(minor)` or `(major)`.
 See `PUBLISH.md` for instructions on how to publish a new version.
 -->
 
+
+## v1.3.3 - 21ac6ca
+
 - (patch) Set aspect-ratio on YouTube/Wistia, responsiveness fixes
 - (patch) Remove minified versions of patched Prism files